+++ src/main/java/com/bolsasdeidead/springboot/backend/apirest/service/AlumnoService.java
@@ -8,6 +8,6 @@ public interface AlumnoService {
 
 	public List<Alumno>findAll();
 	public Alumno save(Alumno alumno);
-	public Alumno findById(int id);
-	public void deleteById(int id);
+	public Alumno findById(Long id);
+	public void deleteById(Long id);
 }
